♻️ refactor: 영역별 event handler 파일 분리
