+++ src/index.js
@@ -1,11 +1,11 @@
 import initDisplayDate from "./js/init/initDisplayDate.js";
-import initEventHandler from "./js/init/initEventHandler.js";
 import initInputBarDate from "./js/init/initInputBarDate.js";
+import initEventHandler from "./js/init/initEventHandler.js";
 
 const init = () => {
-  initEventHandler();
   initDisplayDate();
   initInputBarDate();
+  initEventHandler();
 };
 
 document.addEventListener("DOMContentLoaded", init);
