+++ src/js/events/dateDisplay/index.js
@@ -1,0 +1,12 @@
+import handlePrevButtonClick from "./handlePrevButtonClick.js";
+import handleNextButtonClick from "./handleNextButtonClick.js";
+
+const dateDisplayEventHandler = () => {
+  const $prevButton = document.querySelector(".prev-button");
+  const $nextButton = document.querySelector(".next-button");
+
+  $prevButton.addEventListener("click", handlePrevButtonClick);
+  $nextButton.addEventListener("click", handleNextButtonClick);
+};
+
+export default dateDisplayEventHandler;
