+++ src/js/events/inputBar/handlePriceInputKeyup.js
