+++ src/js/init/initEventHandler.js
@@ -1,20 +1,8 @@
-import handlePrevButtonClick from "../events/handlePrevButtonClick.js";
-import handleNextButtonClick from "../events/handleNextButtonClick.js";
-import handleTypeIconClick from "../events/handleTypeIconClick.js";
-import handlePriceInputKeyup from "../events/handlePriceInputKeyup.js";
+import dateDisplayEventHandler from "../events/dateDisplay/index.js";
 import inputBarEventHandler from "../events/inputBar/index.js";
 
 const initEventHandler = () => {
-  const $prevButton = document.querySelector(".prev-button");
-  const $nextButton = document.querySelector(".next-button");
-  const $typeIcon = document.querySelector(".type-icon");
-  const $priceInput = document.querySelector("#price");
-
-  $prevButton.addEventListener("click", handlePrevButtonClick);
-  $nextButton.addEventListener("click", handleNextButtonClick);
-  $typeIcon.addEventListener("click", handleTypeIconClick);
-  $priceInput.addEventListener("keyup", handlePriceInputKeyup);
-
+  dateDisplayEventHandler();
   inputBarEventHandler();
 };
